+++ main/StraightWind.h
@@ -54,9 +54,9 @@ public:
 		_age = 0;
 	}
 
-	void calculateWind( double tc, double gs, double th, double tas  );
-	static double calculateSpeed( double angle1, double speed1, double angle2, double speed2  );
-	static double calculateAngle( double angle1, double speed1, double angle2, double speed2  );
+	void calculateWind( float tc, float gs, float th, float tas  );
+	static float calculateSpeed( float angle1, float speed1, float angle2, float speed2  );
+	static float calculateAngle( float angle1, float speed1, float angle2, float speed2  );
 	void newCirclingWind( float angle, float speed );
 	void test();
 	int getAge() { return _age; }
@@ -72,13 +72,13 @@ public:
 private:
 
 	int    nunberOfSamples;  // current number of samples
-	double averageTas;             // TAS in km/h
-	double averageTH;          // sum of Compass true heading
-	double averageTC;          // sum of GPS heading (true course)
-	double averageGS;		   // average ground speed
-	double windDir;            // calculated wind direction
-	double windSpeed;          // calculated wind speed in Km/h
-	double lastWindSpeed;          // calculated wind speed in Km/h
+	float averageTas;             // TAS in km/h
+	float averageTH;          // sum of Compass true heading
+	float averageTC;          // sum of GPS heading (true course)
+	float averageGS;		   // average ground speed
+	float windDir;            // calculated wind direction
+	float windSpeed;          // calculated wind speed in Km/h
+	float lastWindSpeed;          // calculated wind speed in Km/h
 	bool   lowAirspeed;
 	float  circlingWindDir;
 	float  circlingWindDirReverse;
